requirements.txt
torch
tqdm
numpy
svg.py